LayoutTests/imported/w3c/web-platform-tests/css/css-tables/visibility-collapse-col-005-expected.txt

Visibility collapse

Spec
Setting a spanning column to visibility:collapse changes table width but not height. The right two columns here have been collapsed.


PASS col visibility:collapse doesn't change table height
FAIL col visibility:collapse changes table width assert_equals: col visibility:collapse changes table width expected 100 but got 111
